Remove get_subject_session_list from pipelines
