+++ clinica/pipelines/t1_freesurfer/t1_freesurfer_pipeline.py
@@ -32,40 +32,6 @@ class T1FreeSurfer(cpe.Pipeline):
             image_ids = extract_image_ids(t1_freesurfer_files)
         return image_ids
 
-    def get_subject_session_list(self, input_dir, tsv_file, is_bids_dir, base_dir):
-        """Parse a BIDS or CAPS directory to get the subjects and sessions."""
-        from colorama import Fore
-        from clinica.utils.stream import cprint
-        from clinica.utils.filemanip import extract_subjects_sessions_from_filename
-        from clinica.utils.inputs import clinica_file_reader
-        from clinica.utils.input_files import T1W_NII
-
-        super(T1FreeSurfer, self).get_subject_session_list(input_dir, tsv_file, is_bids_dir, base_dir)
-
-        if not tsv_file:
-            cprint("Selecting T1w files in BIDS directory")
-            t1w_files = clinica_file_reader(self.subjects, self.sessions,
-                                            self.bids_directory, T1W_NII, False)
-            self.subjects, self.sessions = extract_subjects_sessions_from_filename(t1w_files)
-
-        # Display image(s) already present in CAPS folder
-        output_ids = self.get_processed_images(self.caps_directory, self.subjects, self.sessions)
-        processed_participants, processed_sessions = extract_subjects_sessions_from_filename(output_ids)
-        if len(processed_participants) > 0:
-            cprint("%sClinica found %s image(s) already processed in CAPS directory:%s" %
-                   (Fore.YELLOW, len(processed_participants), Fore.RESET))
-            for p_id, s_id in zip(processed_participants, processed_sessions):
-                cprint("%s\t%s | %s%s" % (Fore.YELLOW, p_id, s_id, Fore.RESET))
-            if self.overwrite_caps:
-                output_folder = "<CAPS>/subjects/sub-<participant_id>/ses-<session_id>/t1/freesurfer_cross_sectional"
-                cprint("%s\nOutput folders in %s will be recreated.\n%s" % (Fore.YELLOW, output_folder, Fore.RESET))
-            else:
-                cprint("%s\nImage(s) will be ignored by Clinica.\n%s" % (Fore.YELLOW, Fore.RESET))
-                input_ids = [p_id + '_' + s_id for p_id, s_id in zip(self.subjects, self.sessions)]
-                processed_ids = [p_id + '_' + s_id for p_id, s_id in zip(processed_participants, processed_sessions)]
-                to_process_ids = list(set(input_ids) - set(processed_ids))
-                self.subjects, self.sessions = extract_subjects_sessions_from_filename(to_process_ids)
-
     def check_custom_dependencies(self):
         """Check dependencies that can not be listed in the `info.json` file."""
         pass
@@ -101,15 +67,37 @@ class T1FreeSurfer(cpe.Pipeline):
             ClinicaBIDSError: If there are duplicated files or missing files for any subject
         """
         import os
+        from colorama import Fore
         import nipype.interfaces.utility as nutil
         import nipype.pipeline.engine as npe
+        from clinica.iotools.utils.data_handling import check_volume_location_in_world_coordinate_system
         from clinica.utils.exceptions import ClinicaBIDSError, ClinicaException
-        from clinica.utils.stream import cprint
-        from clinica.utils.ux import print_images_to_process
+        from clinica.utils.filemanip import extract_subjects_sessions_from_filename, save_participants_sessions
         from clinica.utils.inputs import clinica_file_reader
         from clinica.utils.input_files import T1W_NII
-        from clinica.iotools.utils.data_handling import check_volume_location_in_world_coordinate_system
-        from clinica.utils.filemanip import save_participants_sessions
+        from clinica.utils.stream import cprint
+        from clinica.utils.ux import print_images_to_process
+
+        # Display image(s) already present in CAPS folder
+        # ===============================================
+        output_ids = self.get_processed_images(self.caps_directory, self.subjects, self.sessions)
+        processed_participants, processed_sessions = extract_subjects_sessions_from_filename(output_ids)
+        if len(processed_participants) > 0:
+            cprint("%sClinica found %s image(s) already processed in CAPS directory:%s" %
+                   (Fore.YELLOW, len(processed_participants), Fore.RESET))
+            for p_id, s_id in zip(processed_participants, processed_sessions):
+                cprint("%s\t%s | %s%s" % (Fore.YELLOW, p_id, s_id, Fore.RESET))
+            if self.overwrite_caps:
+                output_folder = "<CAPS>/subjects/sub-<participant_id>/ses-<session_id>/t1/freesurfer_cross_sectional"
+                cprint("%s\nOutput folders in %s will be recreated.\n%s" % (Fore.YELLOW, output_folder, Fore.RESET))
+            else:
+                cprint("%s\nImage(s) will be ignored by Clinica.\n%s" % (Fore.YELLOW, Fore.RESET))
+                input_ids = [p_id + '_' + s_id for p_id, s_id in
+                             zip(self.subjects, self.sessions)]
+                processed_ids = [p_id + '_' + s_id for p_id, s_id in
+                                 zip(processed_participants, processed_sessions)]
+                to_process_ids = list(set(input_ids) - set(processed_ids))
+                self.subjects, self.sessions = extract_subjects_sessions_from_filename(to_process_ids)
 
         # Inputs from anat/ folder
         # ========================
